backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.4.0
